# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.CreateShiftTemplatePayload:
    properties:
      day_of_week:
        maximum: 6
        minimum: 0
        type: integer
      end_time:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      role_ids:
        items:
          type: integer
        type: array
      start_time:
        type: string
    required:
    - end_time
    - name
    - start_time
    type: object
...
  main.UpdateShiftTemplatePayload:
    properties:
      day_of_week:
        maximum: 6
        minimum: 0
        type: integer
      end_time:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      role_ids:
        items:
          type: integer
        type: array
      start_time:
        type: string
    type: object
...
  store.Role:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      restaurant_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  store.ShiftTemplate:
    properties:
      created_at:
        type: string
      day_of_week:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      name:
        type: string
      restaurant_id:
        type: integer
      role_ids:
        description: Stored as JSONB column
        items:
          type: integer
        type: array
      start_time:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Restaurant Employee Scheduling Application (RESA)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RESA API
paths:
...
  /restaurants/{restaurant_id}/shift-templates:
    get:
      consumes:
      - application/json
      description: Fetches all shift templates for a restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.ShiftTemplate'
            type: array
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Lists restaurant's shift templates
      tags:
      - shift-template
    post:
      consumes:
      - application/json
      description: Creates a shift template for a restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Shift template payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateShiftTemplatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.ShiftTemplate'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Creates a shift template
      tags:
      - shift-template
  /restaurants/{restaurant_id}/shift-templates/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a shift template by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Shift Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Deletes a shift template
      tags:
      - shift-template
    get:
      consumes:
      - application/json
      description: Fetches a shift template by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Shift Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.ShiftTemplate'
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetches a shift template
      tags:
      - shift-template
    patch:
      consumes:
      - application/json
      description: Updates a shift template by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Shift Template ID
        in: path
        name: id
        required: true
        type: integer
      - description: Shift template payload with optional fields
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateShiftTemplatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.ShiftTemplate'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Updates a shift template
      tags:
      - shift-template
  /restaurants/{restaurant_id}/shift-templates/{id}/roles:
    get:
      consumes:
      - application/json
      description: Retrieves all roles associated with a shift template
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Shift Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Role'
            type: array
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get roles for a shift template
      tags:
      - shift-template
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"